+++ src/pages/hyeonze/Login/Login.jsx
@@ -15,7 +15,7 @@ const Login = () => {
   // 사용자 데이터 저장 && 버튼활성화
   const [idInput, setIdInput] = useState('');
   const [pwInput, setPwInput] = useState('');
-  const [classOfBtn, setClassOfBtn] = useState('');
+  const [classOfBtn, setClassOfBtn] = useState(false);
   const [isValidatedId, setIsValidatedId] = useState(false);
   const [isValidatedPw, setIsValidatedPw] = useState(false);
 
@@ -24,13 +24,13 @@ const Login = () => {
     idInput.indexOf('@') !== -1
       ? setIsValidatedId(true)
       : setIsValidatedId(false);
-    isValidatedId && isValidatedPw ? setClassOfBtn('blue') : setClassOfBtn('');
+    isValidatedId && isValidatedPw ? setClassOfBtn(true) : setClassOfBtn(false);
   };
 
   const changePwInput = e => {
     setPwInput(e.target.value);
     pwInput.length > 4 ? setIsValidatedPw(true) : setIsValidatedPw(false);
-    isValidatedId && isValidatedPw ? setClassOfBtn('blue') : setClassOfBtn('');
+    isValidatedId && isValidatedPw ? setClassOfBtn(true) : setClassOfBtn(false);
   };
 
   return (
@@ -51,7 +51,7 @@ const Login = () => {
             onChange={changePwInput}
           />
           {/* <Link to="/main">로그인</Link> */}
-          <span className={classOfBtn} onClick={goToMain}>
+          <span className={classOfBtn ? 'blue' : ''} onClick={goToMain}>
             로그인
           </span>
         </div>
