+++ src/pages/hyeonze/Main/Feeds/Comment/Comment.scss
@@ -1,0 +1,48 @@
+span {
+  padding-left: 10px;
+
+  small {
+    color: #a3a3a3;
+  }
+
+  &:first-child {
+    padding: 10px;
+    font-weight: bold;
+  }
+
+  svg.delete {
+    display: none;
+    cursor: pointer;
+  }
+
+  &:hover svg.delete {
+    display: inline-block;
+  }
+
+  svg.hearts {
+    color: #a3a3a3;
+    position: absolute;
+    right: 20px;
+    transition: color 1s;
+    cursor: pointer;
+
+    path {
+      color: #a3a3a3;
+    }
+  }
+
+  svg.red {
+    path {
+      color: #f00;
+    }
+  }
+}
+svg.red {
+  color: #f00;
+}
+
+svg.fa-times-circle {
+  margin-left: 5px;
+  color: #a3a3a3;
+  cursor: pointer;
+}
