+++ src/pages/hyeonze/Main/Feeds/Feeds.scss
@@ -1,0 +1,108 @@
+@use '../../../../styles/mixin.scss' as mixin;
+@use '../../../../styles/color.scss' as clr;
+
+.feeds {
+  position: absolute;
+  top: 77px;
+  width: 60%;
+  background: #fff;
+
+  .top_menu {
+    @include mixin.flex($justify: space-between, $align: center);
+    height: 82px;
+    background: #fff;
+
+    span {
+      @include mixin.flex($align: center);
+      // display: flex;
+      // align-items: center;
+      margin-left: 20px;
+      font-weight: bold;
+
+      &:last-child {
+        margin-right: 22px;
+        color: #2a2a2a;
+      }
+
+      img {
+        width: 42px;
+        height: 42px;
+        border-radius: 50%;
+        margin-right: 20px;
+      }
+    }
+  }
+  article {
+    height: 100%;
+
+    img {
+      width: 100%;
+      height: 818px;
+    }
+
+    .bottom_menu {
+      @include mixin.flex($align: center);
+      // display: flex;
+      position: relative;
+      // align-items: center;
+      width: 100%;
+      height: 62px;
+      background: #fff;
+
+      span > svg {
+        margin: 0 10px;
+        font-size: 32px;
+        color: #414141;
+      }
+
+      span:last-child > svg {
+        position: absolute;
+        top: 15px;
+        right: 0;
+      }
+
+      span:first-child > svg > path {
+        color: #eb4b59;
+      }
+    }
+  }
+
+  .comments {
+    display: flex;
+    flex-direction: column;
+
+    span {
+      padding-left: 10px;
+    }
+  }
+
+  .add_comments {
+    position: absolute;
+    @include mixin.flex($justify: space-between, $align: center);
+    bottom: -62px;
+    width: 100%;
+    height: 62px;
+    background: #fff;
+    border-top: 1px solid clr.$clr-border;
+    font-size: 14px;
+
+    input {
+      width: 90%;
+      height: 50%;
+      margin-left: 15px;
+      border: none;
+    }
+
+    button {
+      margin-right: 15px;
+      border: none;
+      color: clr.$clr-primary;
+      opacity: 0.5;
+      background: #fff;
+
+      &.blue {
+        opacity: 1;
+      }
+    }
+  }
+}
