+++ src/pages/hyeonze/Main/Main.scss
@@ -1,5 +1,4 @@
 @use '../../../styles/mixin.scss' as mixin;
-@use '../../../styles/color.scss' as clr;
 
 #root {
   @include mixin.flex($justify: center, $align: flex-start);
@@ -12,350 +11,6 @@
     position: relative;
     z-index: 10;
     width: 80%;
-
-    .feeds {
-      position: absolute;
-      top: 77px;
-      width: 60%;
-      background: #fff;
-
-      .top_menu {
-        @include mixin.flex($justify: space-between, $align: center);
-        height: 82px;
-        background: #fff;
-
-        span {
-          @include mixin.flex($align: center);
-          // display: flex;
-          // align-items: center;
-          margin-left: 20px;
-          font-weight: bold;
-
-          &:last-child {
-            margin-right: 22px;
-            color: #2a2a2a;
-          }
-
-          img {
-            width: 42px;
-            height: 42px;
-            border-radius: 50%;
-            margin-right: 20px;
-          }
-        }
-      }
-      article {
-        height: 100%;
-
-        img {
-          width: 100%;
-          height: 818px;
-        }
-
-        .bottom_menu {
-          @include mixin.flex($align: center);
-          // display: flex;
-          position: relative;
-          // align-items: center;
-          width: 100%;
-          height: 62px;
-          background: #fff;
-
-          span > svg {
-            margin: 0 10px;
-            font-size: 32px;
-            color: #414141;
-          }
-
-          span:last-child > svg {
-            position: absolute;
-            top: 15px;
-            right: 0;
-          }
-
-          span:first-child > svg > path {
-            color: #eb4b59;
-          }
-        }
-      }
-
-      .comments {
-        display: flex;
-        flex-direction: column;
-
-        span {
-          padding-left: 10px;
-
-          small {
-            color: #a3a3a3;
-          }
-
-          &:first-child {
-            padding: 10px;
-            font-weight: bold;
-          }
-
-          svg.delete {
-            display: none;
-            cursor: pointer;
-          }
-
-          &:hover svg.delete {
-            display: inline-block;
-          }
-
-          svg.hearts {
-            color: #a3a3a3;
-            position: absolute;
-            right: 20px;
-            transition: color 1s;
-            cursor: pointer;
-
-            path {
-              color: #a3a3a3;
-            }
-          }
-        }
-        svg.red {
-          color: #f00;
-        }
-
-        svg.fa-times-circle {
-          margin-left: 5px;
-          color: #a3a3a3;
-          cursor: pointer;
-        }
-      }
-
-      .add_comments {
-        position: absolute;
-        @include mixin.flex($justify: space-between, $align: center);
-        bottom: -62px;
-        width: 100%;
-        height: 62px;
-        background: #fff;
-        border-top: 1px solid clr.$clr-border;
-        font-size: 14px;
-
-        input {
-          width: 90%;
-          height: 50%;
-          margin-left: 15px;
-          border: none;
-        }
-
-        button {
-          margin-right: 15px;
-          border: none;
-          color: clr.$clr-primary;
-          opacity: 0.5;
-          background: #fff;
-
-          &.blue {
-            opacity: 1;
-          }
-        }
-      }
-    }
-
-    .main-right {
-      position: absolute;
-      left: 60%;
-      top: 77px;
-      width: 30%;
-      margin-left: 20px;
-      background: #fafafa;
-
-      img {
-        width: 42px;
-        height: 42px;
-        border-radius: 50%;
-      }
-
-      .stories {
-        position: relative;
-        @include mixin.flex($justify: space-between, $align: flex-start);
-        // display: flex;
-        overflow: hidden;
-        margin-top: 20px;
-        width: 100%;
-        height: 250px;
-        border-radius: 5px;
-        border: 1px solid clr.$clr-border;
-        // justify-content: space-between;
-        background: #fff;
-
-        img {
-          z-index: 2;
-          border: 2px solid #fff;
-        }
-
-        ul li:before {
-          content: '';
-          position: absolute;
-          display: inline-block;
-          z-index: 1;
-          left: 12px;
-          width: 48px;
-          height: 48px;
-          border-radius: 50%;
-          background: clr.$instaGradation;
-        }
-
-        span {
-          padding: 10px;
-        }
-
-        > span:first-child {
-          color: #999;
-        }
-
-        ul {
-          position: absolute;
-          width: 100%;
-          top: 30px;
-
-          li {
-            @include mixin.flex($align: center);
-            position: relative;
-            height: 60px;
-            padding-left: 15px;
-
-            div {
-              position: relative;
-
-              span {
-                position: absolute;
-                width: 100px;
-
-                &:first-child {
-                  top: -25px;
-                  font-weight: bold;
-                }
-
-                &:last-child {
-                  top: -5px;
-                  color: #a3a3a3;
-                }
-              }
-            }
-          }
-        }
-      }
-
-      .user {
-        position: relative;
-
-        span:first-child {
-          @include mixin.flex($justify: center, $align: center);
-          width: 66px;
-          height: 66px;
-          border-radius: 50%;
-          background: #000;
-          color: #fff;
-        }
-
-        span:nth-child(2) {
-          position: absolute;
-          left: 93px;
-          top: 20px;
-          font-weight: bold;
-        }
-
-        span:last-child {
-          position: absolute;
-          left: 93px;
-          top: 40px;
-          color: #9b9b9b;
-        }
-      }
-
-      .recommended {
-        @include mixin.flex($justify: space-between, $align: flex-start);
-        position: relative;
-        // display: flex;
-        overflow: hidden;
-        margin-top: 20px;
-        margin-bottom: 20px;
-        width: 100%;
-        height: 220px;
-        border-radius: 5px;
-        border: 1px solid clr.$clr-border;
-        // justify-content: space-between;
-        background: #fff;
-
-        ul {
-          position: absolute;
-          width: 100%;
-          top: 30px;
-
-          li {
-            @include mixin.flex($justify: flex-start, $align: center);
-            position: relative;
-            // display: flex;
-            height: 60px;
-            padding-left: 15px;
-            // justify-content: flex-start;
-            // align-items: center;
-
-            div {
-              position: relative;
-              width: 100%;
-
-              > span {
-                position: absolute;
-
-                &:first-child {
-                  top: -25px;
-                  font-weight: bold;
-                  width: 140px;
-                }
-
-                &:nth-child(2) {
-                  top: -5px;
-                  color: #a3a3a3;
-                  width: 140px;
-                }
-
-                &:last-child {
-                  right: 0;
-                  top: -15px;
-                  color: clr.$clr-primary;
-                }
-              }
-            }
-          }
-        }
-
-        span {
-          padding: 10px;
-        }
-
-        > span:first-child {
-          color: #999;
-        }
-      }
-
-      .main-right-copy {
-        position: relative;
-
-        > span,
-        > span > span {
-          color: #c7c7c7;
-          font-size: 16px;
-        }
-
-        > span:first-of-type {
-          position: absolute;
-        }
-
-        > span:last-child {
-          position: absolute;
-          top: 70px;
-        }
-      }
-    }
-    // }
   }
 
   @media screen and (max-width: 1000px) {
