+++ src/pages/hyeonze/Main/MainRight/MainRight.scss
@@ -1,0 +1,200 @@
+@use '../../../../styles/mixin.scss' as mixin;
+@use '../../../../styles/color.scss' as clr;
+
+.main-right {
+  position: absolute;
+  left: 60%;
+  top: 77px;
+  width: 30%;
+  margin-left: 20px;
+  background: #fafafa;
+
+  img {
+    width: 42px;
+    height: 42px;
+    border-radius: 50%;
+  }
+
+  .stories {
+    position: relative;
+    @include mixin.flex($justify: space-between, $align: flex-start);
+    // display: flex;
+    overflow: hidden;
+    margin-top: 20px;
+    width: 100%;
+    height: 250px;
+    border-radius: 5px;
+    border: 1px solid clr.$clr-border;
+    // justify-content: space-between;
+    background: #fff;
+
+    img {
+      z-index: 2;
+      border: 2px solid #fff;
+    }
+
+    ul li:before {
+      content: '';
+      position: absolute;
+      display: inline-block;
+      z-index: 1;
+      left: 12px;
+      width: 48px;
+      height: 48px;
+      border-radius: 50%;
+      background: clr.$instaGradation;
+    }
+
+    span {
+      padding: 10px;
+    }
+
+    > span:first-child {
+      color: #999;
+    }
+
+    ul {
+      position: absolute;
+      width: 100%;
+      top: 30px;
+
+      li {
+        @include mixin.flex($align: center);
+        position: relative;
+        height: 60px;
+        padding-left: 15px;
+
+        div {
+          position: relative;
+
+          span {
+            position: absolute;
+            width: 100px;
+
+            &:first-child {
+              top: -25px;
+              font-weight: bold;
+            }
+
+            &:last-child {
+              top: -5px;
+              color: #a3a3a3;
+            }
+          }
+        }
+      }
+    }
+  }
+
+  .user {
+    position: relative;
+
+    span:first-child {
+      @include mixin.flex($justify: center, $align: center);
+      width: 66px;
+      height: 66px;
+      border-radius: 50%;
+      background: #000;
+      color: #fff;
+    }
+
+    span:nth-child(2) {
+      position: absolute;
+      left: 93px;
+      top: 20px;
+      font-weight: bold;
+    }
+
+    span:last-child {
+      position: absolute;
+      left: 93px;
+      top: 40px;
+      color: #9b9b9b;
+    }
+  }
+
+  .recommended {
+    @include mixin.flex($justify: space-between, $align: flex-start);
+    position: relative;
+    // display: flex;
+    overflow: hidden;
+    margin-top: 20px;
+    margin-bottom: 20px;
+    width: 100%;
+    height: 220px;
+    border-radius: 5px;
+    border: 1px solid clr.$clr-border;
+    // justify-content: space-between;
+    background: #fff;
+
+    ul {
+      position: absolute;
+      width: 100%;
+      top: 30px;
+
+      li {
+        @include mixin.flex($justify: flex-start, $align: center);
+        position: relative;
+        // display: flex;
+        height: 60px;
+        padding-left: 15px;
+        // justify-content: flex-start;
+        // align-items: center;
+
+        div {
+          position: relative;
+          width: 100%;
+
+          > span {
+            position: absolute;
+
+            &:first-child {
+              top: -25px;
+              font-weight: bold;
+              width: 140px;
+            }
+
+            &:nth-child(2) {
+              top: -5px;
+              color: #a3a3a3;
+              width: 140px;
+            }
+
+            &:last-child {
+              right: 0;
+              top: -15px;
+              color: clr.$clr-primary;
+            }
+          }
+        }
+      }
+    }
+
+    span {
+      padding: 10px;
+    }
+
+    > span:first-child {
+      color: #999;
+    }
+  }
+
+  .main-right-copy {
+    position: relative;
+
+    > span,
+    > span > span {
+      color: #c7c7c7;
+      font-size: 16px;
+    }
+
+    > span:first-of-type {
+      position: absolute;
+    }
+
+    > span:last-child {
+      position: absolute;
+      top: 70px;
+    }
+  }
+}
